ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
pydantic==2.8.2
python-dotenv==1.0.1
